Unify owned Args types between platforms
