+++ library/std/src/sys/args/wasi.rs
@@ -2,18 +2,14 @@
 
 use crate::ffi::{CStr, OsStr, OsString};
 use crate::os::wasi::ffi::OsStrExt;
-use crate::{fmt, vec};
 
-pub struct Args {
-    iter: vec::IntoIter<OsString>,
-}
-
-impl !Send for Args {}
-impl !Sync for Args {}
+#[path = "common.rs"]
+mod common;
+pub use common::Args;
 
 /// Returns the command line arguments
 pub fn args() -> Args {
-    Args { iter: maybe_args().unwrap_or(Vec::new()).into_iter() }
+    Args::new(maybe_args().unwrap_or(Vec::new()))
 }
 
 fn maybe_args() -> Option<Vec<OsString>> {
@@ -29,33 +25,5 @@ fn maybe_args() -> Option<Vec<OsString>> {
             ret.push(OsStr::from_bytes(s.to_bytes()).to_owned());
         }
         Some(ret)
-    }
-}
-
-impl fmt::Debug for Args {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.iter.as_slice().fmt(f)
-    }
-}
-
-impl Iterator for Args {
-    type Item = OsString;
-    fn next(&mut self) -> Option<OsString> {
-        self.iter.next()
-    }
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.iter.size_hint()
-    }
-}
-
-impl ExactSizeIterator for Args {
-    fn len(&self) -> usize {
-        self.iter.len()
-    }
-}
-
-impl DoubleEndedIterator for Args {
-    fn next_back(&mut self) -> Option<OsString> {
-        self.iter.next_back()
     }
 }
